Refactor UnixTerminal and CygwinTerminal.
