+++ src/main/java/com/googlecode/lanterna/terminal/ansi/CygwinTerminal.java
@@ -22,8 +22,6 @@ import com.googlecode.lanterna.TerminalSize;
 
 import java.io.*;
 import java.nio.charset.Charset;
-import java.util.Timer;
-import java.util.TimerTask;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -33,104 +31,72 @@ import java.util.regex.Pattern;
  *
  * @author Martin
  */
-public class CygwinTerminal extends UnixTerminal {
+public class CygwinTerminal extends UnixishTerminal {
 
     private static final Pattern STTY_SIZE_PATTERN = Pattern.compile(".*rows ([0-9]+);.*columns ([0-9]+);.*");
 
-    private Timer resizeCheckTimer;
+    private static String sttyLocation = findProgram("stty.exe");
 
     public CygwinTerminal(
             InputStream terminalInput,
             OutputStream terminalOutput,
             Charset terminalCharset) throws IOException {
-        super(terminalInput, terminalOutput, terminalCharset);
+        super(terminalInput, terminalOutput, terminalCharset, CtrlCBehaviour.TRAP);
 
         //Make sure to set an initial size
         onResized(80, 24);
-        resizeCheckTimer = null;
+
+        saveSTTY();
+        setCBreak(true);
+        setEcho(false);
+        sttyMinimum1CharacterForRead();
+        setupShutdownHook();
     }
 
     @Override
     public TerminalSize getTerminalSize() {
         try {
-            String stty = exec(findSTTY(), "-F", "/dev/pty0", "-a"); //exec(findShell(), "-c", "echo $PPID");
+            String stty = exec(findSTTY(), "-F", "/dev/pty0", "-a");
             Matcher matcher = STTY_SIZE_PATTERN.matcher(stty);
             if(matcher.matches()) {
                 return new TerminalSize(Integer.parseInt(matcher.group(2)), Integer.parseInt(matcher.group(1)));
             }
             else {
-                return new TerminalSize(80, 20);
+                return new TerminalSize(80, 24);
             }
         }
         catch(Throwable e) {
-            return new TerminalSize(80, 20);
+            return new TerminalSize(80, 24);
         }
     }
 
     @Override
-    public void enterPrivateMode() throws IOException {
-        super.enterPrivateMode();
-        setCBreak(true);
-        setEcho(false);
-        stty1CharacterForRead();
-        resizeCheckTimer = new Timer("CygwinTerminalResizeChecker");
-        resizeCheckTimer.scheduleAtFixedRate(new TimerTask() {
-            @Override
-            public void run() {
-                //queryTerminalSize();
-            }
-        }, 1000, 1000);
-    }
-
-    @Override
-    public void exitPrivateMode() throws IOException {
-        resizeCheckTimer.cancel();
-        setEcho(true);
-        super.exitPrivateMode();
-        setCBreak(false);
-    }
-
-    @Override
-    public void setCBreak(boolean cbreakOn) throws IOException {
-        sttyCBreak(cbreakOn);
-    }
-
-    @Override
-    public void setEcho(boolean echoOn) throws IOException {
-        sttyKeyEcho(echoOn);
-    }
-
-    private static void sttyKeyEcho(final boolean enable) throws IOException {
+    protected void sttyKeyEcho(final boolean enable) throws IOException {
         exec(findSTTY(), "-F", "/dev/pty0", (enable ? "echo" : "-echo"));
-        /*
-         exec(findShell(), "-c",
-         "/bin/stty.exe " + (enable ? "echo" : "-echo") + " < /dev/tty");
-         */
     }
 
-    private static void stty1CharacterForRead() throws IOException {
+    @Override
+    protected void sttyMinimum1CharacterForRead() throws IOException {
         exec(findSTTY(), "-F", "/dev/pty0", "min", "1");
-        /*
-         exec(findShell(), "-c",
-         "/bin/stty.exe min " + nrCharacters + " < /dev/tty");
-         */
     }
 
-    private static void sttyCBreak(final boolean enable) throws IOException {
-        exec(findSTTY(), "-F", "/dev/pty0", (enable ? "cbreak" : "icanon"));
-        /*
-         exec(findShell(), "-c",
-         "/bin/stty.exe " + (enable ? "-icanon" : "icanon") + " < /dev/tty");
-         */
+    @Override
+    protected void sttyICanon(final boolean enable) throws IOException {
+        exec(findSTTY(), "-F", "/dev/pty0", (enable ? "icanon" : "cbreak"));
     }
 
-    @SuppressWarnings("unused")
-    private static String findShell() {
-        return findProgram("sh.exe");
+    @Override
+    protected String sttySave() throws IOException {
+        return exec(findSTTY(), "-F", "/dev/pty0", "-g").trim();
     }
 
-    private static String findSTTY() {
-        return findProgram("stty.exe");
+    @Override
+    protected void sttyRestore(String tok) throws IOException {
+        exec(findSTTY(), "-F", "/dev/pty0", tok);
+    }
+
+    protected String findSTTY() {
+        return sttyLocation;
     }
 
     private static String findProgram(String programName) {
@@ -142,26 +108,6 @@ public class CygwinTerminal extends UnixTerminal {
             }
         }
         return programName;
-    }
-
-    private static String exec(String... cmd) throws IOException {
-        ProcessBuilder pb = new ProcessBuilder(cmd);
-        Process process = pb.start();
-        ByteArrayOutputStream stdoutBuffer = new ByteArrayOutputStream();
-        InputStream stdout = process.getInputStream();
-        int readByte = stdout.read();
-        while(readByte >= 0) {
-            stdoutBuffer.write(readByte);
-            readByte = stdout.read();
-        }
-        ByteArrayInputStream stdoutBufferInputStream = new ByteArrayInputStream(stdoutBuffer.toByteArray());
-        BufferedReader reader = new BufferedReader(new InputStreamReader(stdoutBufferInputStream));
-        StringBuilder builder = new StringBuilder();
-        while(reader.ready()) {
-            builder.append(reader.readLine());
-        }
-        reader.close();
-        return builder.toString();
     }
 
 }
